Surfaces: hide SurfaceError implementation
